// backend/src/questions/dto/create-question.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { QuestionDifficulty } from '@prisma/client';
import {
  ArrayMinSize,
  IsArray,
  IsBoolean,
  IsEnum,
  IsNotEmpty,
  IsOptional,
  IsString,
  IsUUID,
  ValidateNested
} from 'class-validator';
import { Type } from 'class-transformer';
import { CreateQuestionAlternativeDto } from './create-question-alternative.dto';

export class CreateQuestionDto {

  @ApiProperty({ example: 'f8ec8f23-fcbd-49f1-8049-d6f20b94b310' })
  @IsUUID()
  subjectId!: string;

  @ApiProperty({ example: 'What is the primary force responsible for keeping an aircraft in the air?' })
  @IsString()
  @IsNotEmpty()
  statement!: string;

  @ApiProperty({
    required: false,
    example: 'Lift is generated by the pressure difference between the upper and lower surfaces of the wing.'
  })
  @IsOptional()
  @IsString()
  explanation?: string;

  @ApiProperty({ enum: QuestionDifficulty })
  @IsEnum(QuestionDifficulty)
  difficulty!: QuestionDifficulty;

  @ApiProperty({
    required: false,
    default: true
  })
  @IsOptional()
  @IsBoolean()
  isActive?: boolean;

  @ApiProperty({
    type: [CreateQuestionAlternativeDto],
    example: [
      {
        letter: 'A',
        content: 'Lift is generated by pressure difference.',
        isCorrect: true
      },
      {
        letter: 'B',
        content: 'Lift is generated by engine thrust.',
        isCorrect: false
      },
      {
        letter: 'C',
        content: 'Lift is generated by aircraft weight.',
        isCorrect: false
      },
      {
        letter: 'D',
        content: 'Lift is generated by centrifugal force.',
        isCorrect: false
      }
    ]
  })
  @IsArray()
  @ArrayMinSize(4)
  @ValidateNested({ each: true })
  @Type(() => CreateQuestionAlternativeDto)
  alternatives!: CreateQuestionAlternativeDto[];



}